pyproject.toml
[tool.poetry]
name = "elvia_stats"
version = "0.1.0"
description = ""
authors = ["Knut H Pederseen <knuthp@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
requests = "^2.26.0"
python-dotenv = "^0.19.1"
pandas = "^1.3.4"
altair = "^4.1.0"
matplotlib = "^3.4.3"
streamlit = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^3.0.0"
jupyterlab = "^3.2.1"
black = {extras = ["jupyter"], version = "^21.11b0"}
isort = "^5.10.1"
flake8 = "^4.0.1"
requests-mock = "^1.9.3"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"